Rename things in preparation to add another AWS SDK test type
