+++ tests/Agent/IntegrationTests/ContainerApplications/AwsSdkTestApp/AwsSdkExercisers/AwsSdkTestType.cs
@@ -1,7 +1,7 @@
 // Copyright 2020 New Relic, Inc. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-namespace AwsSdkTestApp.AwsSdkExerciser;
+namespace AwsSdkTestApp.AwsSdkExercisers;
 
 public enum AwsSdkTestType
 {
