+++ tests/Agent/IntegrationTests/ContainerApplications/AwsSdkTestApp/SQSBackgroundService/SQSReceiverService.cs
@@ -6,7 +6,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Amazon.SQS.Model;
-using AwsSdkTestApp.AwsSdkExerciser;
+using AwsSdkTestApp.AwsSdkExercisers;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
 using NewRelic.Api.Agent;
@@ -56,7 +56,7 @@ namespace AwsSdkTestApp.SQSBackgroundService
         private async Task<IEnumerable<Message>> ProcessRequestAsync(string queueUrl)
         {
             _logger.LogInformation("Received a request to receive a message from {Queue}", queueUrl);
-            using var awsSdkExerciser = new AwsSdkExerciser.AwsSdkExerciser(AwsSdkTestType.SQS);
+            using var awsSdkExerciser = new AwsSdkExercisers.AwsSdkExerciser(AwsSdkTestType.SQS);
             awsSdkExerciser.SQS_SetQueueUrl(queueUrl);
             _logger.LogInformation("Receiving a message from {Queue}", queueUrl);
             var messages = await awsSdkExerciser.SQS_ReceiveMessageAsync();
